Revert 15 develop

+++ database/migrations/2019_11_11_140652_create_restaurants_table.php
@@ -17,7 +17,7 @@ class CreateRestaurantsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('location');
-            $table->integer('phone');
+            $table->bigInteger('phone');
             $table->string('food_type');
             $table->timestamps();
         });
